pyproject.toml
[project]
name = "dcc-rail-controller"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.2",
    "pyserial>=3.5",
    "simplepyble>=0.12.1",
]
